api/requirements.txt
django==1.11.6
djangorestframework==3.7.1
Pillow==4.3.0
django-ckeditor==5.3.1
psycopg2==2.7.3.2
django-jet==1.0.6
requests==2.18.4
django-analytical==2.4.0
django_cors_headers==2.2.0